fix bug

+++ ping-service/src/main/java/com/freedom/limit/PingProperties.java
@@ -8,13 +8,13 @@ import org.springframework.stereotype.Component;
 @ConfigurationProperties(prefix = "ping")
 public class PingProperties {
 
-    private Double limitQps;
+    private Integer limitQps;
 
-    public Double getLimitQps() {
+    public Integer getLimitQps() {
         return limitQps;
     }
 
-    public void setLimitQps(Double limitQps) {
+    public void setLimitQps(Integer limitQps) {
         this.limitQps = limitQps;
     }
 }
